Adjust pull-bot configuration for new PR template (#23236)
Summary:
The updated Dangerfile config lets pull-bot warn when the PR lacks a Summary, on top of the existing Test Plan and Changelog warnings.

It will also be a bit more lenient when it comes to the changelog, making the `-` optional after TYPE.

Finally, I removed the "large PR" warning. We haven't had a problem with large PRs recently, and I feel we can deal with these on a case by case basis.

[GENERAL] [Changed] Adjust pull-bot to handle new PR template.

Tested against this PR (note the non-standard section titles):

```
$ yarn
$ yarn danger pr https://github.com/facebook/react-native/pull/23236

:clipboard: Missing Summary - <i>Can you add a Summary? To do so, add a "## Summary" section to your PR description. This is a good place to explain the motivation for making this change.</i>
-
:clipboard: Missing Test Plan - <i>Can you add a Test Plan? To do so, add a "## Test Plan" section to your PR description. A Test Plan lets us know how these changes were tested.</i>
-
:clipboard: Missing Changelog - <i>Can you add a Changelog? To do so, add a "## Changelog" section to your PR description. A changelog entry has the following format: [`[CATEGORY] [TYPE] - Message`](http://facebook.github.io/react-native/docs/contributing#changelog).</i>

Found only warnings, not failing the build
✨  Done in 4.86s.
```
Pull Request resolved: https://github.com/facebook/react-native/pull/23236

Differential Revision: D13915716

Pulled By: cpojer

fbshipit-source-id: 6f1c3e67140c14359fc369b36b900aedf9a3bd8a