chore: upgrade `jest-junit` to avoid older jest dependencies (#23476)
Summary:
Upgrade `jest-junit` to the latest version (breaking change between 5 and 6 was reverted). In particular, 6.2.0 removes the dependency on `jest-config` which pulls in `babel-core@6` etc.

I noticed this since `react-native` had `jest-junit` in `dependencies` until #23276. So not as big of an issue. Still a nice cleanup, I'd say

/cc palmerj3

Btw, any reason Jest is still at alpha?

[General] [Fixed] - upgrade `jest-junit` to avoid older jest dependencies
Pull Request resolved: https://github.com/facebook/react-native/pull/23476

Differential Revision: D14100133

Pulled By: cpojer

fbshipit-source-id: 79877261b7ea995ac1f4249a12a6980ec488f655